#1728 Introduced special matchers (abstract and fallback) for PropertyDescriptor properties on masters. If fallback value matcher is sufficient, there is no need to create matcher type from abstract one -- it will be created automatically.
