+++ platform-web-ui/src/main/java/ua/com/fielden/platform/web/view/master/EntityMaster.java
@@ -2,6 +2,10 @@ package ua.com.fielden.platform.web.view.master;
 
 import static java.util.Optional.empty;
 import static java.util.Optional.of;
+import static ua.com.fielden.platform.reflection.AnnotationReflector.getPropertyAnnotation;
+import static ua.com.fielden.platform.reflection.Finder.findFieldByName;
+import static ua.com.fielden.platform.reflection.PropertyTypeDeterminator.determinePropertyType;
+import static ua.com.fielden.platform.utils.EntityUtils.isPropertyDescriptor;
 import static ua.com.fielden.platform.web.centre.EntityCentre.IMPORTS;
 import static ua.com.fielden.platform.web.centre.EntityCentre.createFetchModelForAutocompleterFrom;
 
@@ -10,6 +14,7 @@ import java.util.Optional;
 import com.google.inject.Injector;
 
 import ua.com.fielden.platform.basic.IValueMatcherWithContext;
+import ua.com.fielden.platform.basic.autocompleter.FallbackPropertyDescriptorMatcherWithContext;
 import ua.com.fielden.platform.basic.autocompleter.FallbackValueMatcherWithContext;
 import ua.com.fielden.platform.dao.IEntityDao;
 import ua.com.fielden.platform.dom.DomElement;
@@ -21,12 +26,11 @@ import ua.com.fielden.platform.entity.ActivatableAbstractEntity;
 import ua.com.fielden.platform.entity.DefaultEntityProducerForCompoundMenuItem;
 import ua.com.fielden.platform.entity.DefaultEntityProducerWithContext;
 import ua.com.fielden.platform.entity.IEntityProducer;
+import ua.com.fielden.platform.entity.annotation.IsProperty;
 import ua.com.fielden.platform.entity.annotation.SkipEntityExistsValidation;
 import ua.com.fielden.platform.entity.factory.EntityFactory;
 import ua.com.fielden.platform.entity.factory.ICompanionObjectFinder;
 import ua.com.fielden.platform.entity.query.fluent.fetch;
-import ua.com.fielden.platform.reflection.Finder;
-import ua.com.fielden.platform.reflection.PropertyTypeDeterminator;
 import ua.com.fielden.platform.utils.ResourceLoader;
 import ua.com.fielden.platform.web.centre.EntityCentre;
 import ua.com.fielden.platform.web.centre.api.actions.EntityActionConfig;
@@ -204,15 +208,17 @@ public class EntityMaster<T extends AbstractEntity<?>> implements IRenderable {
             final ICompanionObjectFinder coFinder) {
         
         final boolean isEntityItself = "".equals(propertyName); // empty property means "entity itself"
-        final Class<V> propertyType = (Class<V>) (isEntityItself ? entityType : PropertyTypeDeterminator.determinePropertyType(entityType, propertyName));
-        final IEntityDao<V> co = coFinder.find(propertyType);
-
-        // filtering out of inactive should only happen for activatable properties without SkipEntityExistsValidation present
-        final boolean activeOnly = 
-                ActivatableAbstractEntity.class.isAssignableFrom(propertyType) &&
-                !Finder.findFieldByName(entityType, propertyName).isAnnotationPresent(SkipEntityExistsValidation.class);
-        
-        return new FallbackValueMatcherWithContext<>(co, activeOnly);
+        final Class<V> propertyType = (Class<V>) (isEntityItself ? entityType : determinePropertyType(entityType, propertyName));
+        if (isPropertyDescriptor(propertyType)) {
+            return (IValueMatcherWithContext<T, V>) new FallbackPropertyDescriptorMatcherWithContext<>((Class<AbstractEntity<?>>) getPropertyAnnotation(IsProperty.class, entityType, propertyName).value());
+        } else {
+            final IEntityDao<V> co = coFinder.find(propertyType);
+            // filtering out of inactive should only happen for activatable properties without SkipEntityExistsValidation present
+            final boolean activeOnly = 
+                    ActivatableAbstractEntity.class.isAssignableFrom(propertyType) &&
+                    !findFieldByName(entityType, propertyName).isAnnotationPresent(SkipEntityExistsValidation.class);
+            return new FallbackValueMatcherWithContext<>(co, activeOnly);
+        }
     }
 
     @Override
